+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [0.2.0] - 2018-03-28
+
+### Improvements
+
+* Moved `*Task` classes to `lib/thermite/rails/tasks/`.
+* Added rake task for running `cargo update` for all crates.
+
 ## [0.1.0] - 2018-03-27
 
 ### Improvements